+++ addOns/wappalyzer/src/main/java/org/zaproxy/zap/extension/wappalyzer/ExtensionWappalyzer.java
@@ -51,8 +51,8 @@ import org.parosproxy.paros.extension.SessionChangedListener;
 import org.parosproxy.paros.model.Session;
 import org.parosproxy.paros.model.SiteNode;
 import org.parosproxy.paros.view.View;
+import org.zaproxy.addon.pscan.ExtensionPassiveScan2;
 import org.zaproxy.zap.extension.alert.ExampleAlertProvider;
-import org.zaproxy.zap.extension.pscan.ExtensionPassiveScan;
 import org.zaproxy.zap.extension.search.ExtensionSearch;
 import org.zaproxy.zap.utils.ThreadUtils;
 import org.zaproxy.zap.view.ScanPanel;
@@ -87,7 +87,7 @@ public class ExtensionWappalyzer extends ExtensionAdaptor
 
     /** The dependencies of the extension. */
     private static final List<Class<? extends Extension>> EXTENSION_DEPENDENCIES =
-            List.of(ExtensionPassiveScan.class);
+            List.of(ExtensionPassiveScan2.class);
 
     private TechPassiveScanner passiveScanner;
 
@@ -188,12 +188,16 @@ public class ExtensionWappalyzer extends ExtensionAdaptor
         extensionHook.addApiImplementor(new TechApi(this));
         extensionHook.addOptionsParamSet(techDetectParam);
 
-        ExtensionPassiveScan extPScan =
+        getPscanExtension().getPassiveScannersManager().add(passiveScanner);
+        extensionHook.addOptionsChangedListener(passiveScanner);
+    }
+
+    private ExtensionPassiveScan2 getPscanExtension() {
+        ExtensionPassiveScan2 extPScan =
                 Control.getSingleton()
                         .getExtensionLoader()
-                        .getExtension(ExtensionPassiveScan.class);
-        extPScan.addPassiveScanner(passiveScanner);
-        extensionHook.addOptionsChangedListener(passiveScanner);
+                        .getExtension(ExtensionPassiveScan2.class);
+        return extPScan;
     }
 
     @Override
@@ -247,11 +251,7 @@ public class ExtensionWappalyzer extends ExtensionAdaptor
     public void unload() {
         super.unload();
 
-        ExtensionPassiveScan extPScan =
-                Control.getSingleton()
-                        .getExtensionLoader()
-                        .getExtension(ExtensionPassiveScan.class);
-        extPScan.removePassiveScanner(passiveScanner);
+        getPscanExtension().getPassiveScannersManager().remove(passiveScanner);
     }
 
     @Override