+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Depend on Passive Scanner add-on (Issue 7959).
 
 ## [0.16.0] - 2024-11-06
 ### Fixed